Move habitat to v2 plugin.
