+++ .rubocop.yml
@@ -7,6 +7,7 @@ AllCops:
   - 'test/**/*'
   - 'examples/**/*'
   - 'vendor/**/*'
+  - 'lib/plugins/inspec-*/test/**/*'
   - 'lib/bundles/inspec-init/templates/**/*'
   - 'www/demo/**/*'
 AlignParameters:
