--- lib/bundles/inspec-habitat/log.rb
@@ -1,10 +1,0 @@
-# encoding: utf-8
-# author: Adam Leff
-
-require 'mixlib/log'
-
-module Habitat
-  class Log
-    extend Mixlib::Log
-  end
-end
